% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_density_calculation.R
\name{bulk_density}
\alias{bulk_density}
\title{Calculate soil bulk density}
\usage{
bulk_density(
  dataset,
  core_diameter,
  depth_upper,
  depth_lower,
  coarse_fragment_weight,
  soil_dry_weight
)
}
\arguments{
\item{core_diameter}{the diameter of the core used to take the soil sample (units: cm)}

\item{depth_upper}{the value of the upper soil depth increment (units: cm)}

\item{depth_lower}{the value of the lower soil depth increment (units: cm)}

\item{coarse_fragment_weight}{coarse fragment weight, which includes weight of worms (units: g)}

\item{soil_dry_weight}{the oven dry weight of the total soil sample (units: g)}
}
\value{
bulk_density the calculated soil bulk density (units: g cm^-3)
}
\description{
Calculate soil bulk density
}
