% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySE.R
\name{summarySE}
\alias{summarySE}
\title{Calculate summary statistics based on groups}
\usage{
summarySE(
  data = NULL,
  measurevar,
  groupvars = NULL,
  na.rm = FALSE,
  conf.interval = 0.95,
  .drop = TRUE
)
}
\arguments{
\item{data}{the dataset containing the parameters}

\item{measurevar}{the variable that was measured}

\item{groupvars}{the grouped variables or treatment; can provide a list}

\item{na.rm}{remove NA values; default is FALSE}

\item{conf.interval}{the confidence intervals; default is 95\%}

\item{sample_wet_weight}{the weight of the field moist soil sample and the bag if applicable (units: g)}
}
\value{
datac dataframe with standard deviation, standard error, and confidence intervals
}
\description{
Calculate summary statistics based on groups
}
